Fix login redirect cookie to work in iframes
Browsers are adding new default restrictions to cookies when a page is loaded within an iframe. By default they will block all cookies if the parent page and iframe source are on different sites. This means that if a Mendix app is loaded in an externally hosted iframe (e.g. a dashboard page) this cookie will be blocked and the login redirect will not work. By adding the `SameSite=None` and `Secure` (when applicable) options the restriction is lifted and the cookie will not be blocked.