fix(pluggable-widgets-tools): make start:js and start:ts a noop
It is needed because in the past `npm start` for web widgets executed start:ts and
start:server concurrently, but starting with 8.10 it started to call start:server only. So
from the two of them one had to become a noop (otherwise we break <8.10 widgets).
It makes sense to pick the former, since that breaks less important `dev` command
for newer widgets.